backend/requirements.txt
# django==4.2
# djangorestframework==3.14
# gunicorn==20.1.0
# pymongo==4.3.3
# numpy==1.23.5
# scikit-learn==1.2.0
# tensorflow==2.14.1
# Pillow==9.5.0
# matplotlib==3.7.1  
  
django==4.2
djangorestframework==3.14
gunicorn==20.1.0
pymongo==4.3.3
numpy==1.23.5
scikit-learn==1.2.0
tensorflow-macos==2.13.0
Pillow==9.5.0
matplotlib==3.7.1
typing-extensions==4.5.0  # Remplacez par la version compatible
